Extract CallNumbers::OtherShelfkey for consistency across call number types
